+++ frontend/next-homework/src/app/components/button/index.tsx
@@ -2,7 +2,7 @@
 
 import Image from "next/image";
 import styles from "./styles.module.css";
-import add from "../../../public/icons/add.svg";
+import add from "../../../../public/icons/add.svg";
 
 // 게시글 등록 페이지 버튼
 export const addressSearchButton = () => {
@@ -22,6 +22,14 @@ export const postSubmitButton = ({ onClick }, buttonActiveStyle: boolean) => {
       style={{ backgroundColor: buttonActiveStyle ? "var(--n-main, #2974e5)" : "var(--gray-300, #c7c7c7)" }}
     >
       등록하기
+    </button>
+  );
+};
+
+export const postEditButton = ({ onClick }) => {
+  return (
+    <button id="postEditButton" className={`${styles.check} ${styles.submit}`} onClick={onClick}>
+      수정하기
     </button>
   );
 };
