+++ frontend/next-homework/src/app/components/button/styles.module.css
@@ -67,4 +67,19 @@
 }
 
 @media (min-width: 781px) {
+  .uploadImg {
+    width: 16rem;
+    height: 16rem;
+  }
+
+  .check {
+    width: auto;
+
+    font-size: 1.8rem;
+    line-height: 2.4rem;
+  }
+
+  .submit {
+    color: #fff;
+  }
 }
